util: Accept NULL pointers for pid and bundlePath
This gives callers a clear way to express disinterest in either of the
possible extracted values.  This commit also adjusts kill_container
and delete_container to not ask for bundlePath, which they didn't
need.  That saves a bit of time spent parsing JSON and a bit of memory
spent holding the unused value.